Fix the temporal attention calculation.
For time step t, we will use the spatial embedding from
time 0:t-1 as key and value and the mask token as query
and apply directional attention. That embedding will be
actual positional emebdding of that agent at time t. We
will apply spatial attention at time t to get the spat-
-ial embedding at t. This step will keep going on for
the next steps.

Signed-off-by: Yash Ranjan <yashranjan@ufl.edu>